requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.7.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloup==3.0.5
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docutils==0.19
executing==2.0.1
fastapi==0.109.2
fastjsonschema==2.19.1
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.5.4
generativeai==0.0.1
gitdb==4.0.11
GitPython==3.1.41
glcontext==2.5.0
google==3.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==1.6.7
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-generativeai==0.7.1
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.63.0
grpcio-status==1.62.2
grpcio-tools==1.63.0
gspread==5.12.4
gspread-dataframe==3.3.1
gTTS==2.5.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.6
ifaddr==0.2.0
imbalanced-learn==0.12.3
imblearn==0.0
importlib-metadata==6.11.0
intel-openmp==2021.4.0
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
isosurfaces==0.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
manim==0.18.1
ManimPango==0.5.0
mapbox-earcut==1.0.1
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
ml-dtypes==0.3.2
moderngl==5.10.0
moderngl-window==2.4.6
mpmath==1.3.0
multidict==6.0.5
multipledispatch==1.0.0
namex==0.0.8
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.9
networkx==3.3
nicegui==1.4.28
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.6
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
param==2.1.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
plotly==5.22.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.23.4
pscript==0.7.7
psutil==5.9.7
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.26.1
pycparser==2.21
pydantic==2.8.0
pydantic_core==2.20.0
pydeck==0.8.1b0
pydub==0.25.1
pyglet==2.0.15
Pygments==2.17.2
pyparsing==3.1.1
pyrr==0.10.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.0
scipy==1.11.4
screeninfo==0.8.1
seaborn==0.13.1
Send2Trash==1.8.2
setuptools==69.0.3
simple-websocket==1.0.0
six==1.16.0
skia-pathops==0.8.0.post1
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
srt==3.5.3
stability-sdk==0.8.6
stack-data==0.6.3
starlette==0.36.3
streamlit==1.36.0
svgelements==1.9.6
sympy==1.12.1
tbb==2021.13.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-intel==2.16.2
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
transformers==4.42.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uri-template==1.3.0
uritemplate==3.0.1
urllib3==2.1.0
uvicorn==0.30.1
validators==0.22.0
vbuild==0.8.2
watchdog==3.0.0
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.9
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.0
yarl==1.9.4
zipp==3.17.0
